techdebt: [PL-29832]: TaskDataV2 move out of 960-persistence (#40256)
* Update cron expression to run every hour

* Moved TaskDataV2 out of 960-persistence to 950-delegate-tasks.
